@@ -5,6 +5,6 @@ resolver = "2"
 [workspace.package]
 version = "0.4.0"
 authors = ["Ethan Roseman <ethteck@gmail.com>"]
-edition = "2021"
+edition = "2024"
 license = "MIT OR Apache-2.0"
 repository = "https://github.com/etheteck/coddog"
